pyproject.toml
[project]
name = "orbit-console-nextjs-tests"
version = "1.0.0"
description = "E2E browser + API tests for orbit_console_nextjs (all Python)"
requires-python = ">=3.11"
dependencies = []

[project.optional-dependencies]
dev = [
    "httpx>=0.24.0",
    "pytest>=7.0.0",
    "pytest-playwright>=0.5.0",
    "playwright>=1.52.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
]

[tool.pytest.ini_options]
testpaths = ["e2e", "tests"]
python_files = ["test_*.py"]
addopts = "--strict-markers --strict-config -v"
markers = [
    "e2e: Playwright browser tests against CONSOLE_URL",
    "console_api: API route tests against CONSOLE_URL",
]
